+++ RollTemplate/Program.cs
@@ -15,7 +15,7 @@ namespace RollTemplate
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new Main());
         }
     }
 }
